==> python/original_sentiment/1_sentiment_analysis.ipynb <==
# Import dependencies
import pandas as pd
import numpy as np
from vaderSentiment.vaderSentiment import SentimentIntensityAnalyzer
analyzer = SentimentIntensityAnalyzer()

# Read in the Scripts CSV file
scripts_df = pd.read_csv('../rawData/scripts.csv')
#scripts_df.head()

# View dimensions of dataset
scripts_df.shape

# Define list of main characters
primaryChars = ["JERRY", "GEORGE", "KRAMER", "ELAINE"]

# Setting the cut-off for our secondary characters at 150 lines of dialogue
secondaryChars = ["NEWMAN", "MORTY", "HELEN", "FRANK","SUSAN", "ESTELLE", "PUDDY"]

# Add column for Sentiment Score
scripts_df = scripts_df.reindex(columns = np.append(scripts_df.columns.values, ['Sentiment']))
scripts_df.head()

# Create blank coolumns for positive, negative and compound sentiment episodes
scripts_df["PosSent"] = np.repeat(0, scripts_df.shape[0])
scripts_df["NegSent"] = np.repeat(0, scripts_df.shape[0])
scripts_df["CompSent"] = np.repeat(0, scripts_df.shape[0])

scripts_df.head()

# Loop to re-name Character column and add Sentiment Score

for i in range(scripts_df.shape[0]):

    #########################
    # Re-assign character names into "OTHER" or "SECONDARY" for non-main characters
    #########################
    
    char = scripts_df["Character"][i]

    if (char not in primaryChars) & (char not in secondaryChars):
        scripts_df["Character"][i] = "OTHER"

    elif char not in primaryChars:
        scripts_df["Character"][i] = "SECONDARY"


    #########################
    # Sentiment Analysis
    #########################
    try:
        # Setting each line of dialogue as a string, otherwise numbers give us errors
        scripts_df["Sentiment"][i] = analyzer.polarity_scores(str(scripts_df["Dialogue"][i]))["compound"]
    
    except:
        # Also using try / except to limit that
        scripts_df["Sentiment"][i] = 0

    #########################
    # Print notifications to make sure the code is running
    #########################

    if i % 2000 == 0:
        perc_complete = round(i / scripts_df.shape[0] * 100, 2)
        print("Percent Complete: " + str(perc_complete) + "%")
    
    if i == scripts_df.shape[0]:
        print("-----------------------------------")
        print("Loop Complete!")
        

# Save scripts_df to a new CSV file, so that we don't have to do the loop again
scripts_df.to_csv('../rawData/cleaned_scripts.csv')
