==> python/original_sentiment/3_final_merge.ipynb <==
# Import dependencies
import pandas as pd
import numpy as np

# Read in the CSV files
merged_df = pd.read_csv('../rawData/merged_scripts.csv')
episode_df = pd.read_csv('../rawData/episode_info.csv')

# Drop the "Unnamed: 0" and "Unnamed: 0.1" columns
merged_df = merged_df.drop("Unnamed: 0", 1)
episode_df = episode_df.drop("Unnamed: 0", 1)

# Drop S01E00, since we don't have any sentiment for that
episode_df = episode_df[1:episode_df.shape[0]]

merged_df.tail()

episode_df.head()

### Merge both data frames on SEID
seinfeld_df = pd.merge(episode_df, merged_df, on = "SEID", how = "left")
seinfeld_df.head()

# Drop EpisodeNo_y and Season_y
seinfeld_df = seinfeld_df.drop("EpisodeNo_y", 1)
seinfeld_df = seinfeld_df.drop("Season_y", 1)

# Rename EpsodeNo_x and Season_x
seinfeld_df = seinfeld_df.rename(index=str, columns = {"EpisodeNo_x": "EpisodeNo", "Season_x": "Season"})
seinfeld_df.head()

# Export to CSV
seinfeld_df.to_csv('../rawData/seinfeld.csv')

##### Not doing anything with it yet, but something to consider is how to handle the two-part episodes...

# There are 3 two-part episodes... we're going to divide those into two separate episodes for the purpose of our data
# Two-part episodes are:
# 1) S01E01
# 2) S05E18
# 3) S09E23

import warnings
warnings.filterwarnings('ignore')

s0101 = seinfeld_df.loc[seinfeld_df["SEID"] == "S01E01",:]

#Halve the LineCount and WordCount variables
s0101["LineCountTotal"] = s0101["LineCountTotal"] / 2
s0101["WordCountTotal"] = s0101["WordCountTotal"] / 2

s0101["LineCountJerry"] = s0101["LineCountJerry"] / 2
s0101["WordCountJerry"] = s0101["WordCountJerry"] / 2

s0101["LineCountGeorge"] = s0101["LineCountGeorge"] / 2
s0101["WordCountGeorge"] = s0101["WordCountGeorge"] / 2

s0101["LineCountKramer"] = s0101["LineCountKramer"] / 2
s0101["WordCountKramer"] = s0101["WordCountKramer"] / 2

s0101["LineCountElaine"] = s0101["LineCountElaine"] / 2
s0101["WordCountElaine"] = s0101["WordCountElaine"] / 2

s0101["LineCountMainChars"] = s0101["LineCountMainChars"] / 2
s0101["WordCountMainChars"] = s0101["WordCountMainChars"] / 2

s0101["LineCountSecondary"] = s0101["LineCountSecondary"] / 2
s0101["WordCountSecondary"] = s0101["WordCountSecondary"] / 2

s0101["LineCountOther"] = s0101["LineCountOther"] / 2
s0101["WordCountOther"] = s0101["WordCountOther"] / 2

s0101

s518 = seinfeld_df.loc[seinfeld_df["SEID"] == "S05E18",:]

#Halve the LineCount and WordCount variables
s518["LineCountTotal"] = s518["LineCountTotal"] / 2
s518["WordCountTotal"] = s518["WordCountTotal"] / 2

s518["LineCountJerry"] = s518["LineCountJerry"] / 2
s518["WordCountJerry"] = s518["WordCountJerry"] / 2

s518["LineCountGeorge"] = s518["LineCountGeorge"] / 2
s518["WordCountGeorge"] = s518["WordCountGeorge"] / 2

s518["LineCountKramer"] = s518["LineCountKramer"] / 2
s518["WordCountKramer"] = s518["WordCountKramer"] / 2

s518["LineCountElaine"] = s518["LineCountElaine"] / 2
s518["WordCountElaine"] = s518["WordCountElaine"] / 2

s518["LineCountMainChars"] = s518["LineCountMainChars"] / 2
s518["WordCountMainChars"] = s518["WordCountMainChars"] / 2

s518["LineCountSecondary"] = s518["LineCountSecondary"] / 2
s518["WordCountSecondary"] = s518["WordCountSecondary"] / 2

s518["LineCountOther"] = s518["LineCountOther"] / 2
s518["WordCountOther"] = s518["WordCountOther"] / 2

s518

s0923 = seinfeld_df.loc[seinfeld_df["SEID"] == "S09E23",:]

# Halve the LineCount and WordCount variables
s0923["LineCountTotal"] = s0923["LineCountTotal"] / 2
s0923["WordCountTotal"] = s0923["WordCountTotal"] / 2

s0923["LineCountJerry"] = s0923["LineCountJerry"] / 2
s0923["WordCountJerry"] = s0923["WordCountJerry"] / 2

s0923["LineCountGeorge"] = s0923["LineCountGeorge"] / 2
s0923["WordCountGeorge"] = s0923["WordCountGeorge"] / 2

s0923["LineCountKramer"] = s0923["LineCountKramer"] / 2
s0923["WordCountKramer"] = s0923["WordCountKramer"] / 2

s0923["LineCountElaine"] = s0923["LineCountElaine"] / 2
s0923["WordCountElaine"] = s0923["WordCountElaine"] / 2

s0923["LineCountMainChars"] = s0923["LineCountMainChars"] / 2
s0923["WordCountMainChars"] = s0923["WordCountMainChars"] / 2

s0923["LineCountSecondary"] = s0923["LineCountSecondary"] / 2
s0923["WordCountSecondary"] = s0923["WordCountSecondary"] / 2

s0923["LineCountOther"] = s0923["LineCountOther"] / 2
s0923["WordCountOther"] = s0923["WordCountOther"] / 2

# Create a new row for each half of episode
s09231 = s0923
s09231["EpisodeNo"] = 23.1
s09231["SEID"] = "S09E23.1"

s09232 = s0923
s09232["EpisodeNo"] = 23.1
s09232["SEID"] = "S09E23.1"
